pyproject.toml
[project]
name = "cleanfolder"
version = "0.1.0"
description = "LLM-powered folder cleaning agent for macOS — detect duplicates, categorize files, and reclaim storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.0",
    "click>=8.3.2",
    "httpx>=0.28.1",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
    "send2trash>=2.1.0",
    "thefuzz>=0.22.1",
]

[project.scripts]
cleanfolder = "cleanfolder.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
